Update won function in xo.cpp
Optimized sub-optimal function that checks if the game has ended.
It before checked if all the boxes on a certain line were equal to either 'X' or 'O' depending on the case and they only need to be equal to each other. Since it checks the game every time a move has been made, it means it is certain that if the game is now over the player who won is the one who moved last.